backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Redis & Celery
redis==5.0.1
celery==5.3.6

# OpenAI
openai==1.10.0

# Export
python-docx==1.1.0
reportlab==4.0.9
ebooklib==0.18
markdown==3.5.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
pydantic[email]==2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
